pyproject.toml
[tool.poetry]
name = "web_crawl_screenshot"
version = "0.1.0"
description = "A tool to crawl websites like a human, take screenshots, build site structure maps, and diff against sitemap.xml"
authors = ["Adam Breen <adam@adambreen.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://adambreen.com"
repository = "https://github.com/adambreen/web_crawl_screenshot"
documentation = "https://github.com/adambreen/web_crawl_screenshot/blob/main/README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
playwright = "^1.39.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
pyyaml = "^6.0"
lxml = "^4.9.3"

[tool.poetry.dev-dependencies]
pytest = "^7.3.0"
requests-mock = "^1.11.0"
black = "^23.3.0"
ruff = "^0.0.281"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
